+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### next release (8.8.1)
 
+- enforce the WMS layer limit
 - [The next improvement]
 
 #### 8.8.0 - 2025-02-18